ctf/web/HTBank/web_htbank/challenge/flask_frontend/requirements.txt
Flask
flask_mysqldb
pyjwt
mysqlclient
requests